// types/type.ts
export type ActiveElement = {
  name: string;
  value: string;
  icon: string;
} | null;
// ...
export type NavbarProps = {
  activeElement: ActiveElement;
  imageInputRef: React.MutableRefObject<HTMLInputElement | null>;
  handleImageUpload: (e: React.ChangeEvent<HTMLInputElement>) => void;
  handleActiveElement: (element: ActiveElement) => void;
  onToggleLeftSidebar?: () => void;
  onToggleRightSidebar?: () => void;
};